+++ packages/web/src/layout/components/AppMain.vue
@@ -1,0 +1,5 @@
+<script setup lang="ts"></script>
+
+<template>
+  <main class="content"></main>
+</template>
